utils: take care of vendor/model ID mismatch in both root/unit directories
In the case that the content includes the different value of vendor/model
ID fields in root/unit directories, in modalias the ones in root directory
is prior to the ones in unit directory, even if the modalias should refer
to the data of unit directory.

This commit fixes the bug.